regsensitivity: regression sensitivity analysis

Implements the identified-set and breakdown-point sensitivity analyses of Diegert, Masten and Poirier (2026), Oster (2019), and Masten and Poirier (2026) for the coefficient of interest in a linear regression with omitted variables.

Core functions

regsen_bounds() – identified set across sensitivity parameters

regsen_breakdown() – smallest sensitivity value at which a hypothesis fails

regsen_summary() – default summary sweep (DMP + Oster)

plot.regsensitivity() – ggplot2 visualization

Reference data

bfg2020 – subset of the Bazzi, Fiszbein & Gebresilasse (2020) replication data, used to demonstrate the package.

See also

Author

Maintainer: Mike Nguyen nguyennghia1301@gmail.com [copyright holder]